+++ README.md
@@ -1,15 +1,17 @@
 ## 기능 요구 사항
 ### API
-[ ] 사용자가 자신의 이름으로 본인의 예약 목록을 조회할 수 있다
-    [ ] 전체 예약 조회는 관리자만 가능하다
-[ ] 사용자가 본인의 예약을 취소할 수 있다
-[ ] 사용자가 본인 예약의 날짜/시간을 변경할 수 있다
+
+- [ ]  사용자가 자신의 이름으로 본인의 예약 목록을 조회할 수 있다
+    - [x]  전체 예약 조회는 관리자만 가능하다
+- [ ]  사용자가 본인의 예약을 취소할 수 있다
+- [ ]  사용자가 본인 예약의 날짜/시간을 변경할 수 있다
 
 ### 예외 처리
-[ ] 지나간 날짜/시간에 대한 예약 생성은 불가능하다
-[ ] 같은 날짜+시간+테마에 이미 예약이 있으면 중복 예약을 거부한다
-[ ] 예약이 존재하는 시간을 삭제할 수 없다
-[ ] 유효하지 않은 입력값을 거부한다
+
+- [ ]  지나간 날짜/시간에 대한 예약 생성은 불가능하다
+- [ ]  같은 날짜+시간+테마에 이미 예약이 있으면 중복 예약을 거부한다
+- [ ]  예약이 존재하는 시간을 삭제할 수 없다
+- [ ]  유효하지 않은 입력값을 거부한다
 
 ## API 명세
