+++ src/main/java/roomescape/reservation/controller/AdminReservationController.java
@@ -1,0 +1,30 @@
+package roomescape.reservation.controller;
+
+import java.util.List;
+import java.util.stream.Collectors;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.dto.ReservationsResponse;
+import roomescape.reservation.service.ReservationService;
+
+@RestController
+@RequestMapping("/admin")
+public class AdminReservationController {
+
+    private final ReservationService reservationService;
+
+    public AdminReservationController(ReservationService reservationService) {
+        this.reservationService = reservationService;
+    }
+
+    @GetMapping("/reservations")
+    public ResponseEntity<ReservationsResponse> readAll() {
+        List<ReservationResponse> reservations = reservationService.findAll().stream()
+                .map(ReservationResponse::from)
+                .collect(Collectors.toList());
+        return ResponseEntity.ok().body(new ReservationsResponse(reservations, reservations.size()));
+    }
+}
